Clean up dead code and extract mobile customizer sub-components
Remove unused COLOR_TO_KEY map, wire PlayersPage grid to the
already-built MemoizedCricketCard, and extract SnapStrip and
ColorEditorPanel out of CustomizerMobile to shrink it and reuse the
scroll-snap pattern. Move hexToHSL/hslToHex to lib/color.ts as the
shared source and to avoid a circular import between the two new
components.
